% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{addCellProp}
\alias{addCellProp}
\title{Add cell properties}
\usage{
addCellProp(vat, cell.prop, prop.name, meta.key = NULL, meta.value = NULL)
}
\arguments{
\item{vat}{vat entity}

\item{cell.prop}{cell property's values}

\item{prop.name}{cell property's name}

\item{meta.key}{cell property's meta feature key}

\item{meta.value}{property's meta feature value}
}
\description{
Add cell properties
}
