% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{addGeneProp}
\alias{addGeneProp}
\title{Add gene properties}
\usage{
addGeneProp(vat, gene.prop, prop.name, meta.key = NULL, meta.value = NULL)
}
\arguments{
\item{vat}{vat entity}

\item{gene.prop}{gene property's values}

\item{prop.name}{gene property's name}

\item{meta.key}{property's meta feature key}

\item{meta.value}{property's meta feature value}
}
\description{
Add gene properties
}
