% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{addPropMeta}
\alias{addPropMeta}
\title{Add meta feature}
\usage{
addPropMeta(prop.meta, meta.key, meta.value)
}
\arguments{
\item{prop.meta}{property name}

\item{meta.key}{meta feature key}

\item{meta.value}{meta feature value}
}
\description{
Add meta feature
}
