% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{buildKNNGraph}
\alias{buildKNNGraph}
\title{Building kNN Graph for clustering}
\usage{
buildKNNGraph(cell.dist, k, dist.type, verbose = TRUE)
}
\arguments{
\item{cell.dist}{cell distant matrix}

\item{k}{kNN's k value}

\item{dist.type}{distanct type}

\item{verbose}{whether or not printing verbose, default}
}
\description{
Building kNN Graph for clustering
}
