% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterGraph}
\alias{clusterGraph}
\title{Clustering Graph}
\usage{
clusterGraph(graph, graph.type = "knn", community.detect = "infomap",
  k = 0)
}
\arguments{
\item{graph}{kNN graph}

\item{graph.type}{can be jaccard, invlogweighted or dice. Default knn}

\item{community.detect}{can be louvain, infomap or markov. Default infomap}

\item{k}{k-value}
}
\description{
Clustering Graph
}
