% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiating.R
\name{doDiffAnalysis}
\alias{doDiffAnalysis}
\title{Differential Analysis for group1 and group2 (according to group.key)}
\usage{
doDiffAnalysis(vat, group1, group2 = NULL, group.key = "cluster",
  use.genes = NULL, method = "t", min.logfc = 0.25,
  only.pos = FALSE, min.avg = 0.1, min.diff.avg = -Inf,
  top.num = Inf, min.cells = 1, verbose = TRUE)
}
\arguments{
\item{vat}{VAT entity}

\item{group1}{Group ID for the first group from vat@cell.props[group.key]}

\item{group2}{Group ID for the second group from vat@cell.props[group.key]
If NULL (default), use all other cells for comparison.}

\item{group.key}{Key value storing Group from colnames(vat@cell.props) for Differential Analysis}

\item{use.genes}{Genes ,Default is to use all genes}

\item{method}{Method which does differential analysise. Available options are:
\itemize{
 \item{"t"} :  t-test (default)
 \item{"wilcox"} : Wilcoxon rank sum test
 \item{"DESeq2} : DE based on a model using the negative binomial distribution (Love et al, Genome Biology, 2014), required DESeq2 library.
}}

\item{min.logfc}{At least X-fold difference (log-scale) between the two groups of cells. Default is 0.25}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{min.avg}{only compare genes which average min.avg cells in either of  two groups, Default is 0.1}

\item{min.diff.avg}{only  compare genes which minimum difference between  two groups. Default is -Inf}

\item{top.num}{only return top.num results. Default is Inf, and return all results}

\item{min.cells}{Minimum number of cells expressing the gene in at least min.cells. Default is 1}

\item{verbose}{Show the progress bar}
}
\value{
p-value adjustment is performed using bonferroni correction based on the total number of genes in the dataset.
}
\description{
Differential Analysis for group1 and group2 (according to group.key)
}
