% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{doPCA}
\alias{doPCA}
\title{do PCA using irlba library}
\usage{
doPCA(vat, pc.num = 50, use.genes = NULL, ...)
}
\arguments{
\item{vat}{vat entity}

\item{pc.num}{default 50}

\item{use.genes}{run PCA using the genes. if NULL, using vat@use.genes. default NULL}
}
\description{
do PCA using irlba library
}
