% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{doTSNE}
\alias{doTSNE}
\title{do tSNE using Rtsne library}
\usage{
doTSNE(vat, dims = 2, analysis.key = "PCA", use.col = 50, ...)
}
\arguments{
\item{vat}{vat entity}

\item{dims}{tSNE dimension}
}
\description{
do tSNE using Rtsne library
}
