% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiating.R
\name{doTTest}
\alias{doTTest}
\title{do t-test for differential analysis}
\usage{
doTTest(vat, cells1, cells2, use.genes, verbose = TRUE)
}
\arguments{
\item{vat}{vat entity}

\item{cells1}{the first group index}

\item{cells2}{the second group index}

\item{use.genes}{the genes for  differential analysis}

\item{verbose}{whether or not showing verbose, default TRUE}
}
\value{
p.value for test, rownames are use.genes
}
\description{
do t-test for differential analysis
}
