% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells}
\usage{
filterCells(vat, prop.name, lower.limit, upper.limit)
}
\arguments{
\item{vat}{VAT Entity}

\item{prop.name}{filtered cell's prop.name}

\item{lower.limit}{filtered lower limit, only keeping the cells which properties are more than lower.limit}

\item{upper.limit}{filtered upper limit, only keeping the cells which properties are less than upper.limit}
}
\value{
filtered data
}
\description{
Filter cells
}
\examples{
 #filtering vat where 0< gene.nums < Inf and 0 < umi < 10000
 vat <- filterCells(vat, prop.name = c("gene.nums","umi"),lower.limit = c(0, 0), upper.limit = c(Inf, 10000))
}
