% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter genes}
\usage{
filterGenes(vat, prop.name, lower.limit, upper.limit)
}
\arguments{
\item{vat}{VAT Entity}

\item{prop.name}{filtered gene's prop.name}

\item{lower.limit}{filtered lower limit, only keeping the genes which properties are more than lower.limit}

\item{upper.limit}{filtered upper limit, only keeping the genes which properties are less than upper.limit}
}
\value{
filtered data
}
\description{
Filter genes
}
\examples{
 #filtering vat where 3 < cell.nums < Inf
 vat <- filterCells(vat, prop.name = c("cell.nums"),lower.limit = c(3), upper.limit = c(Inf))
}
