% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getAnalysisColName}
\alias{getAnalysisColName}
\title{Get the analysis key's colnames}
\usage{
getAnalysisColName(key = "tSNE", dims = c(1:2))
}
\arguments{
\item{key}{analysis key}

\item{dims}{colnames subscript}
}
\description{
Get the analysis key's colnames
}
