% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getAnalysisColnames}
\alias{getAnalysisColnames}
\title{get availbale analysis data colnames}
\usage{
getAnalysisColnames(vat, key = "tSNE")
}
\arguments{
\item{vat}{vat entity}

\item{key}{analysis key, default tSNE}
}
\description{
get availbale analysis data colnames
}
