% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getAnalysisData}
\alias{getAnalysisData}
\title{Get analysis result}
\usage{
getAnalysisData(vat, key = "PCA", cols = c(1:50), as.data.frame = FALSE)
}
\arguments{
\item{vat}{vat entity}

\item{key}{analysis key}

\item{cols}{analysis data column vector}

\item{as.data.frame}{convert to data.frame if TRUE, default FALSE}
}
\description{
Get analysis result
}
