% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getAnalysisList}
\alias{getAnalysisList}
\title{Get analysis list}
\usage{
getAnalysisList(vat, key = "PCA")
}
\arguments{
\item{vat}{vat entity}

\item{key}{analysis key}

\item{analysis}{list}
}
\description{
Get analysis list
}
