% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getCellPropData}
\alias{getCellPropData}
\title{Get cell properties}
\usage{
getCellPropData(vat, prop.name)
}
\arguments{
\item{vat}{vat entity}

\item{prop.name}{cell property's name}
}
\description{
Get cell properties
}
