% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getGenePropData}
\alias{getGenePropData}
\title{Get gene properties}
\usage{
getGenePropData(vat, prop.name)
}
\arguments{
\item{vat}{vat entity}

\item{prop.name}{gene property's name}
}
\description{
Get gene properties
}
