% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getPropMeta}
\alias{getPropMeta}
\title{Get the property meta features}
\usage{
getPropMeta(vat, meta.key)
}
\arguments{
\item{vat}{vat entity}

\item{meta.key}{meta feature key}
}
\description{
Get the property meta features
}
