% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{getUseData}
\alias{getUseData}
\title{Get dataset}
\usage{
getUseData(vat, use.genes = NULL, use.raw = FALSE, drop = FALSE)
}
\arguments{
\item{vat}{vat entity}

\item{use.genes}{use gene names}

\item{use.raw}{return raw data if TRUE, default FALSE}

\item{drop}{whether or not drop}
}
\description{
Get dataset
}
