% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{initVATEntity}
\alias{initVATEntity}
\title{initialize VAT Entity}
\usage{
initVATEntity(data.raw, title = "VAT", min.genes = 0, min.cells = 0,
  log.trans = TRUE, pseudocount = 1, cell.scale = FALSE,
  scale.factor = 10000, verbose = TRUE)
}
\arguments{
\item{data.raw}{data matrix}

\item{title}{dataset name}
}
\description{
initialize VAT Entity
}
