% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load10XH5}
\alias{load10XH5}
\title{load data from H5 file}
\usage{
load10XH5(filename, genome = NULL)
}
\arguments{
\item{filename}{h5 file name for 10X}

\item{genome}{h5group genome name, default NULL, returns first group}
}
\description{
load data from H5 file
}
\examples{
data <- loadH5Data("./filtered_gene_bc_matrices_h5.h5",genome="mm10")
}
