% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load10XPath}
\alias{load10XPath}
\title{load data from 10x path (including barcodes.tsv, genes.tsv and matrix.mtx)}
\usage{
load10XPath(path)
}
\arguments{
\item{path}{directory stroed in 10x data}
}
\description{
load data from 10x path (including barcodes.tsv, genes.tsv and matrix.mtx)
}
\examples{
data <- loadH5Data("./filtered_gene_bc_matrices/mm10")
}
