% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadAnalysis}
\alias{loadAnalysis}
\title{load analysis result manually}
\usage{
loadAnalysis(vat, analysis.data, dims = c(1:2), key = "tSNE")
}
\arguments{
\item{vat}{VATEntity object}

\item{analysis.data}{analysis data for cells}

\item{dims}{the dim values of analysis result}

\item{key}{key value in the anlaysis list}
}
\value{
new vat entity including new analysis result
}
\description{
load analysis result manually
}
