% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadAnalysisFromCSV}
\alias{loadAnalysisFromCSV}
\title{load analysis result manually from csv}
\usage{
loadAnalysisFromCSV(vat, filename, ndims = 2, key = "tSNE", ...)
}
\arguments{
\item{vat}{VATEntity object}

\item{filename}{csv filename}

\item{ndims}{the dim of analysis result}

\item{key}{key value in the anlaysis list}
}
\value{
new vat entity including new analysis result
}
\description{
load analysis result manually from csv
}
