% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadCSVData}
\alias{loadCSVData}
\title{load data from csv file}
\usage{
loadCSVData(filename, gene.names = NULL, cell.names = NULL,
  gene.in.row = TRUE, to.sparseMatrix = TRUE, ...)
}
\arguments{
\item{filename}{csv file names}

\item{gene.names}{set rownames=gene.names, if null, using the first column as gene names. default null}

\item{cell.names}{set colnames=cell.names, if null, using the header as cell names. default null}

\item{gene.in.row}{gene store in row, default TRUE}

\item{to.sparseMatrix}{if TRUE, saving data as sparseMatrix, otherwise matrix. default TRUE}

\item{...}{see also read.csv}
}
\value{
data matrix or sparseMatrix data
}
\description{
load data from csv file
}
\examples{
data <- loadCSVData("./filename.csv")
}
