% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadH5Data}
\alias{loadH5Data}
\title{load data from H5 file}
\usage{
loadH5Data(filename, genome = NULL)
}
\arguments{
\item{filename}{h5 file name}

\item{genome}{h5group genome name, default NULL, returns first group}
}
\value{
data sparseMatrix format
}
\description{
load data from H5 file
}
