% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotAnalysis}
\alias{plotAnalysis}
\title{plot using analysis data}
\usage{
plotAnalysis(vat, dims = c(1:2), title = NULL, key = "tSNE",
  color.data = NULL, gradient = FALSE, colors = NULL, source = "src",
  ...)
}
\arguments{
\item{vat}{vat entity}

\item{dims}{data analysis, dimension index, default c(1:2)}

\item{title}{plot title, default NULL, name as vat@title, if "", no title}

\item{key}{analysis data key value}

\item{color.data}{group data for color}

\item{gradient}{default FALSE}

\item{colors}{color palette, if no color.data, uses colors[1]}

\item{source}{plot id}
}
\value{
p
}
\description{
plot using analysis data
}
