% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGene}
\alias{plotGene}
\title{plot gene expression based on the analysis result}
\usage{
plotGene(vat, genes, dims = c(1, 2), key = "tSNE", gradient = TRUE,
  colors = c("lightgrey", "blue"), ...)
}
\arguments{
\item{dims}{length(dim)=2, 3 (2D or 3D)}

\item{key}{analysis data key}
}
\description{
plot gene expression based on the analysis result
}
