% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMultipleGenes}
\alias{plotMultipleGenes}
\title{plot some gene expressions based on the analysis result}
\usage{
plotMultipleGenes(vat, genes, nrows = 2, dims = c(1, 2), key = "tSNE",
  gradient = TRUE, colors = c("lightgrey", "blue"), ...)
}
\arguments{
\item{vat}{vat entity}

\item{genes}{gene's name for plot}

\item{nrows}{number of rows for plot}

\item{dims}{length(dim)=2, 3 (2D or 3D)}

\item{key}{analysis data key}

\item{gradient}{default FALSE}

\item{colors}{color palette}
}
\description{
plot some gene expressions based on the analysis result
}
