% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPC}
\alias{plotPC}
\title{plot PC maps
#'}
\usage{
plotPC(vat, x.pc, y.pc, z.pc = NULL, title = NULL, color = "black",
  source = "pca", ...)
}
\arguments{
\item{vat}{VAT Entity}

\item{x.pc}{X-axis's PC}

\item{y.pc}{X-axis's PC}

\item{z.pc}{Z-axis's PC}

\item{title}{plot's title}

\item{source}{plot's data source for shinyui}

\item{...}{other parameters for plot_ly}

\item{colors}{color palette}
}
\value{
plot.ly object
}
\description{
plot PC maps
#'
}
