% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPCASDev}
\alias{plotPCASDev}
\title{Plot some analysis data's Standard Deviation (such as PCA)}
\usage{
plotPCASDev(vat, key = "PCA", ndims = 50, ...)
}
\arguments{
\item{vat}{VATEntity}

\item{key}{analyais.key, default PCA}

\item{ndims}{show pc components}
}
\value{
plot object
}
\description{
Plot some analysis data's Standard Deviation (such as PCA)
}
