% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotThreeGenes}
\alias{plotThreeGenes}
\title{plot three gene expression based on the analysis result at a map}
\usage{
plotThreeGenes(vat, gene1, gene2, gene3, dims = c(1, 2), key = "tSNE",
  colors = c("lightgrey", "blue", "orange", "green", "brown", "purple",
  "yellow", "red"), ...)
}
\arguments{
\item{vat}{vat entity}

\item{gene1}{the first gene name for plot}

\item{gene2}{the second gene name for plot}

\item{gene3}{the third gene name for plot}

\item{dims}{length(dim)=2, 3 (2D or 3D)}

\item{key}{analysis data key}

\item{colors}{color palette (more colors??8 colors)}
}
\description{
plot three gene expression based on the analysis result at a map
}
