% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTSNE}
\alias{plotTSNE}
\title{plot tsne maps}
\usage{
plotTSNE(vat, title = NULL, group.id = NULL, gradient = FALSE,
  colors = NULL, source = "tsne", ...)
}
\arguments{
\item{vat}{VAT Entity}

\item{title}{plot's title}

\item{group.id}{cell property for group}

\item{gradient}{whether or not with gradient}

\item{colors}{color palette}

\item{source}{plot's data source for shinyui}

\item{...}{other parameters for plot_ly}
}
\value{
plot.ly object
}
\description{
plot tsne maps
}
\examples{
plotTSNE(vat, title="Example", group.id="cluster", gradient=FALSE)
}
