% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{setGenePercent}
\alias{setGenePercent}
\title{Calculating the expressiong's percent for some pattern's genes}
\usage{
setGenePercent(vat, gene.pattern, prop.name)
}
\arguments{
\item{gene.pattern}{pattern expression for matching gene names}
}
\description{
Calculating the expressiong's percent for some pattern's genes
}
