% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vatentity.R
\docType{methods}
\name{show}
\alias{show}
\title{show method for VATEntity}
\usage{
\S4method{show}{VATEntity}(object)
}
\arguments{
\item{object}{A VATEntity object}
}
\description{
show method for VATEntity
}
