% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainui.R
\name{startVATUI}
\alias{startVATUI}
\title{RNA Graph Group main UI function}
\usage{
startVATUI(data_name)
}
\arguments{
\item{data}{...}
}
\value{

}
\description{
start shiny app UI for Graph Group
}
\examples{

}
\author{
thbin
}
