% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vatentity.R
\docType{class}
\name{VATEntity}
\alias{VATEntity}
\title{The VAT Entity Class}
\description{
The VAT entity is the core data structure for VAT pipeline.
It stores all contents including data, annotations, analysis, etc.
}
\section{Slots}{

\describe{
\item{\code{}}{}
}}

