% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doCluster}
\alias{doCluster}
\title{do Unsupervised Clustering using kNN-graph}
\usage{
doCluster(vat, analysis.key = "PCA", pc.num = 50,
  cluster.name = "cluster", k = 100, dist.type = "euclidean",
  save.KNN = TRUE, verbose = TRUE, ...)
}
\description{
do Unsupervised Clustering using kNN-graph
}
