% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadCSVData}
\alias{loadCSVData}
\title{load data from csv file}
\usage{
loadCSVData(filename, gene.index = 1, cell.index = 1,
  gene.names = NULL, cell.names = NULL, gene.in.row = TRUE, ...)
}
\arguments{
\item{filename}{csv file names}

\item{gene.index}{gene name's subscript}

\item{cell.index}{cell name's subscript}

\item{gene.names}{set rownames=gene.names, default NULL, using the gene.index}

\item{cell.names}{set colnames=cell.names, default NULL, using the cell.index}

\item{gene.in.row}{gene store in row, default TRUE}

\item{...}{see also read.csv}
}
\value{
data sparseMatrix data
}
\description{
load data from csv file
}
\examples{
data <- loadCSVData("./filename.csv")

data <- loadCSVData("./filename.csv", gene.in.row = FALSE)
}
