% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTSNE}
\alias{plotTSNE}
\title{plot tsne maps}
\source{

}
\usage{
plotTSNE(vat, title = NULL, group.id = NULL, gradient = FALSE,
  colors = NULL, source = "tsne", ...)
}
\arguments{
\item{group.id}{cell property for group}
}
\description{
plot tsne maps
}
