% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing.R
\name{setGenePercent}
\alias{setGenePercent}
\title{Calculating the expression's percent for some pattern's genes}
\usage{
setGenePercent(vat, gene.pattern, prop.name)
}
\arguments{
\item{vat}{vat entity}

\item{gene.pattern}{pattern expression for matching gene names}

\item{prop.name}{the saved column name of cell.props}
}
\description{
Calculating the expression's percent for some pattern's genes
}
\examples{
vat <- setGenePercent(vat, gene.pattern="^mt-", prop.name="mt.percent")
}
