% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vatentity.R
\docType{class}
\name{VATEntity}
\alias{VATEntity}
\title{The VAT Entity Class}
\description{
The VAT entity is the core data structure for VAT pipeline.
It stores all contents including data, annotations, analysis, etc.
}
\section{Slots}{

\describe{
\item{\code{title}}{VAT entity's title}

\item{\code{data}}{The normalized expression matrix}

\item{\code{data.raw}}{The raw expression matrix}

\item{\code{gene.props}}{Gene's Properties}

\item{\code{cell.props}}{Cell's Properties}

\item{\code{prop.meta}}{Property's meta features}

\item{\code{use.genes}}{used genes for analysis}

\item{\code{analysis}}{Analysis data list}
}}

